move admin template cmds -> pkg/oc/admin
