+++ pkg/oc/admin/create_bootstrap_project_template.go
