+++ pkg/oc/experimental/ipfailover/keepalived/generator.go
@@ -9,7 +9,7 @@ import (
 
 	dapi "github.com/openshift/origin/pkg/apps/apis/apps"
 	"github.com/openshift/origin/pkg/generate/app"
-	"github.com/openshift/origin/pkg/ipfailover"
+	"github.com/openshift/origin/pkg/oc/experimental/ipfailover/ipfailover"
 )
 
 const defaultInterface = "eth0"
